Create ipfsbed.is-a.dev
For my IPFS image bed (https://ipfsbed.is-an.org).
I am applying for this domain to deploy an IPFS gateway, and I am applying for NS records to facilitate instant management of other records, such as applying for certificates, changing subdomains, etc. I will not use it for any other purpose.